Devdreamsolution/issue11

+++ app/Http/Controllers/Portals/RoleController.php
@@ -39,13 +39,17 @@ class RoleController extends Controller
                 })
                 ->addColumn('action', function ($row) {
                     $btn = '<a href="'. route('portal.usermanage.roles.show', $row->id) .'" data-id="'.$row->id.'" class="btn btn-success btn-sm mb-1 mr-1"><i class="far fa-eye"></i></a>';
-                    $btn .= '<a href="'. route('portal.usermanage.roles.edit', $row->id) .'" data-id="'.$row->id.'" class="btn btn-primary btn-sm mb-1"><i class="far fa-edit"></i></a>';
-                    $btn .= ' <button onclick="deleteData('. "'$row->id'" .')" data-id="'.$row->id.'" class="btn btn-danger btn-sm mb-1"><i class="far fa-trash-alt"></i></button>';
-                    $btn .= '<form id="deleteForm'. $row->id .'" action="'. route('portal.usermanage.roles.destroy', $row->id) .'" method="POST" style="display: none">
-                    <input type="hidden" name="_token" value="'. csrf_token() .'">
-                    <input type="hidden" name="_method" value="DELETE">
-                    @method("DELETE")
-                    </form>';
+                    if (auth()->user()->hasRole('SuperAdmin') || auth()->user()->hasPermissionTo('editrole')) {
+                        $btn .= '<a href="'. route('portal.usermanage.roles.edit', $row->id) .'" data-id="'.$row->id.'" class="btn btn-primary btn-sm mb-1"><i class="far fa-edit"></i></a>';
+                    }
+                    if (auth()->user()->hasRole('SuperAdmin') || auth()->user()->hasPermissionTo('deleterole')) {
+                        $btn .= ' <button onclick="deleteData('. "'$row->id'" .')" data-id="'.$row->id.'" class="btn btn-danger btn-sm mb-1"><i class="far fa-trash-alt"></i></button>';
+                        $btn .= '<form id="deleteForm'. $row->id .'" action="'. route('portal.usermanage.roles.destroy', $row->id) .'" method="POST" style="display: none">
+                        <input type="hidden" name="_token" value="'. csrf_token() .'">
+                        <input type="hidden" name="_method" value="DELETE">
+                        @method("DELETE")
+                        </form>';
+                    }
 
                     return $btn;
                 })
